pyproject.toml
[tool.poetry]
name = "timecard-backend"
version = "0.1.0"
description = ""
authors = ["Gabriel Rebessi <grebessi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.5"
djangorestframework = "^3.11.0"
markdown = "^3.2.1"
django-filter = "^2.2.0"
djangorestframework-simplejwt = {git = "https://github.com/SimpleJWT/django-rest-framework-simplejwt", rev = "master"}
drf-yasg = "^1.17.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
